Add files via upload
This project demonstrates a medical recommendation system implemented  using Python and machine learning techniques, along with web development tools like CSS, HTML, and Flask. The system effectively utilizes a machine learning model to analyze user-provided symptoms and predict potential diseases. It then recommends relevant medications, precautions, and dietary advice for the predicted diseases. The Flask framework facilitates the creation of a user-friendly web interface, allowing users to easily input their symptoms and receive tailored recommendations. This system has the potential to assist users in making informed decisions about their health and well-being, especially in situations where immediate medical attention may not be accessible. However, it is important to note that this system is 
intended for informational purposes only and should not be used as a 
substitute for professional medical advice.